#106 hotfix: 지역 목록 삭제시 기본지역 설정하도록 수정 (#107)
* #106 hotfix : 지역 목록 삭제시 대표지역 설정 추가 , 구조 리팩토링

* #106 test : 지역 목록 삭제시 대표지역 설정 테스트, 구조 리팩토링

* #106 refactor : 내 지역 삭제 리팩토링
