+++ secondhand/src/test/java/com/codesquad/secondhand/domain/member/controller/MemberControllerTest.java
@@ -15,7 +15,6 @@ import com.codesquad.secondhand.BaseControllerTest;
 import com.codesquad.secondhand.annotation.ControllerIntegrationTest;
 import com.codesquad.secondhand.domain.jwt.dto.request.ReissueTokenRequest;
 import com.codesquad.secondhand.domain.jwt.entity.Token;
-import com.codesquad.secondhand.domain.member.dto.request.RegionRequest;
 import com.codesquad.secondhand.domain.member.dto.request.SignupRequest;
 import com.codesquad.secondhand.domain.member.entity.Member;
 
@@ -51,74 +50,6 @@ class MemberControllerTest extends BaseControllerTest {
 			.andExpect(status().isOk());
 
 		Assertions.assertThat(redisUtil.hasKeyBlackList(jwt.getAccessToken())).isTrue();
-	}
-
-	@Test
-	@DisplayName("회원의 아이디와 지역 아이디를 입력받아 회원의 지역을 추가 할 수 있다.")
-	void addRegion() throws Exception {
-		// when & then
-		RegionRequest requestDto = new RegionRequest(3L);
-		String request = objectMapper.writeValueAsString(requestDto);
-		mockMvc.perform(MockMvcRequestBuilders.post("/api/members/1/regions")
-				.header(AUTHORIZATION, JWT_TOKEN_PREFIX + jwt.getAccessToken())
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(request))
-			.andExpect(status().isCreated());
-	}
-
-	@Test
-	@DisplayName("회원의 아이디와 지역 아이디를 입력받아 회원의 지역을 삭제 할 수 있다.")
-	void deleteRegion() throws Exception {
-		// when & then
-		Long memberId = 1L;
-		RegionRequest requestDto = new RegionRequest(4L);
-		memberService.addRegion(memberId, requestDto);
-		String request = objectMapper.writeValueAsString(requestDto);
-		mockMvc.perform(MockMvcRequestBuilders.delete("/api/members/1/regions")
-				.header(AUTHORIZATION, JWT_TOKEN_PREFIX + jwt.getAccessToken())
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(request))
-			.andExpect(status().isNoContent());
-	}
-
-	@Test
-	@DisplayName("회원이 등록한 지역들에서 대표 지역을 설정 할 수 있다.")
-	void setRegion() throws Exception {
-		// when & then
-		Long memberId = 1L;
-		RegionRequest requestSaveDto1 = new RegionRequest(3L);
-		RegionRequest requestSaveDto2 = new RegionRequest(4L);
-		memberService.addRegion(memberId, requestSaveDto1);
-		memberService.addRegion(memberId, requestSaveDto2);
-		RegionRequest requestDto = new RegionRequest(3L);
-		String request = objectMapper.writeValueAsString(requestDto);
-		mockMvc.perform(MockMvcRequestBuilders.put("/api/members/1/regions")
-				.header(AUTHORIZATION, JWT_TOKEN_PREFIX + jwt.getAccessToken())
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(request))
-			.andExpect(status().isNoContent());
-	}
-
-	@Test
-	@DisplayName("회원이 등록한 지역들과 대표지역을 출력한다.")
-	void getSelectedRegionAndRegions() throws Exception {
-		// when & then
-		Long memberId = 1L;
-		// 나의 지역 저장
-		RegionRequest requestSaveDto1 = new RegionRequest(3L);
-		RegionRequest requestSaveDto2 = new RegionRequest(4L);
-		memberService.addRegion(memberId, requestSaveDto1);
-		memberService.addRegion(memberId, requestSaveDto2);
-		// 나의 대표 지역 설정
-		RegionRequest setRegionDto = new RegionRequest(4L);
-		memberService.updateSelectedRegion(memberId, setRegionDto);
-
-		mockMvc.perform(MockMvcRequestBuilders.get("/api/members/1/regions")
-				.header(AUTHORIZATION, JWT_TOKEN_PREFIX + jwt.getAccessToken())
-				.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(jsonPath("$.selectedRegionId").exists())
-			.andExpect(jsonPath("$.regions").exists())
-			.andExpect(status().isOk());
 	}
 
 	private static SignupRequest DummySignUpRequest() {
